add(ipos): move `getDirectoryCID` to `helia` and make `hashOf` util
