+++ services/ipos/package.json
@@ -25,6 +25,7 @@
     "hono": "^4.4.12",
     "ipfs-only-hash": "^4.0.0",
     "ipfs-unixfs-importer": "^15.2.5",
+    "multiformats": "^13.1.3",
     "valibot": "^0.30.0"
   }
 }
